+++ plugins/ccdock/test/test_helper.rb
@@ -34,7 +34,6 @@ module DockTestHelpers
   DEFAULT_CONFIG = {
     'project_name' => 'demoapp',
     'base_host' => 'dev.localhost',
-    'image_name' => 'demoapp',
     'db_name' => 'demoapp_development',
     'db_user' => 'postgres',
     'db_service' => 'postgres'